Now Livestreams are fixed, and count of viewers is displayed.
It was not possible to play a live stream using MPlayer. To
address this, the underlying yt-dlp command was modified when
clicking on a live stream.
   - This can be parameterized in fltube.conf, if desired.

Images were added to display that a video is a livestream, and
another image when no thumbnail is available (for some reason).

Now you can see the count of views (or viewers for a livestream)
in a abbreviated way (1K = 1000, 1M = 1.000.000).