Added ConfigurationManager class and keyboard shortcuts.
The logic of managing the configuration stuff (at fltube.conf)
was extracted to it own class: ConfigurationManager.

A KeyboardShortcuts class was created to manage the various
shortcuts added to the app. This shortcuts can be overwrited by
the user as custom shortcuts. The last can be defined at fltube.conf.

Besides, when writing an input search term, you can press Enter
as a shortcut instead of pressing the "search" button.